src/main/java/racingcar/domain/Car.java
package racingcar.domain;

public class Car {
    private final String name;
    private int position;

    public Car(String name) {
        this.name = name;
        this.position = 0;
    }

    public void goForward(boolean  canGo) {
        if (canGo) {
            position += 1;
        }
    }

    public boolean isAlsoWinner(Car winner) {
        return position == winner.position;
    }

    public String getName() {
        return name;
    }

    public int getPosition() {
        return position;
    }

    @Override
    public String toString() {
        return "Car{" +
                "name='" + name + '\'' +
                ", position=" + position +
                '}';
    }
}
